// src/lib/IubStructure/models/IInputLine.ts
export type InputLine = {
  CodRoteiro: number;
  NomeRoteiro: string;
  ItemPaiRoteiro: string;
  ItemPaiIub: string;
  DescricaoItem: string;
  ItemPaiTipoProduto: string;
  DepositoDestino: string;
  TipoDocumento: string;
  TipoLinha: string;
  Recurso: string;
  Insumo: string;
  InsumoQuantidade: number;
  InsumoDescricao: string;
  InsumoTipo: string;
  InsumoUnidade: string;
  InsumoTipoProduto: string;
  InsumoIndustrializado: string;
  RecursoTempo: string;
  RecursoNome: string;
  RecursoCustoHora: string;
  InsumoCodContaGerencial: string;
  InsumoContaGerencial: string;
};