// src/lib/IubStructure/models/IStructLine.ts
export type StructLine = {
  id: number
  level: number
  NomeRoteiro: string
  ItemPaiRoteiro: string
  ItemPaiIub: string
  DescricaoItem: string
  ItemPaiTipoProduto: string
  TipoLinha: string
  Recurso: string
  Insumo: string
  InsumoQuantidade: number
  InsumoCusto: number
  InsumoPreco: number
  InsumoDescricao: string
  InsumoTipo: string
  InsumoUnidade: string
  InsumoTipoProduto: string
  InsumoIndustrializado: string
  RecursoTempo: string
  RecursoNome: string
  RecursoCustoHora: string
}